Add files via upload
basically it is a personal wordrobe website but we were only able to make it's front UI and we also tried it's it's backend but it shows some errors . so we ended up here but we wish to continue with this project. This website allow the person to take photo and upload it in it and ai will analyse the skin texture,skin tone, body weight, height , face shape of the person , also hairstyle , the ai will suggest fashion sense, combos,accessories,hats etc. which will suit on that person also it also had an personal space in which the pperson can keep their products or combos which they liked,also which they want to wear in future and also which they want to buy . there are some ideas which i want to implement on it but it would take time to implement ....